+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Ampache:
   * Option to change the session timeout with the `config.php` key `music.ampache_session_expiry_time`
     [#1134](https://github.com/owncloud/music/issues/1134)
-  * Support for the action `user`
+  * Support for the actions `search`, `user`, `user_playlists`, `user_smartlists`
 
 ### Changed
 - Ampache: